flask_book_recommendation/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Caching==2.1.0
PyMySQL==1.1.1

# Machine Learning & Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0
